Hotfixes for 1005, 2013 & 2013bis route security. Updated style of pages accordingly to reflect user rights on UI. Addded Watchdog to update sorties states. Removed unwanted console.logs.

+++ src/Controller/VilleController.php
@@ -91,6 +91,8 @@ class VilleController extends AbstractController
             $entityManager->remove($ville);
             $entityManager->flush();
             $this->addFlash('success', 'La ville "'. $name .'" a bien été supprimée !');
+        } else {
+            $this->addFlash('error', 'Vous n`avez pas le droit de faire cela !');
         }
 
         return $this->redirectToRoute('app_ville_index', [], Response::HTTP_SEE_OTHER);
@@ -99,11 +101,13 @@ class VilleController extends AbstractController
     #[Route('/delete/{id}', name: 'app_ville_delete_by_id', methods: ['POST'])]
     public function deleteById(Request $request, Ville $ville, EntityManagerInterface $entityManager): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$ville->getId(), $request->request->get('_token'))) {
+        if ($this->isCsrfTokenValid('delete'.$ville->getId(), $request->request->get('_token')) && $ville->getLieux()[0] == null) {
             $name = $ville->getNom();
             $entityManager->remove($ville);
             $entityManager->flush();
             $this->addFlash('success', 'La ville "'. $name .'" a bien été supprimée !');
+        } else {
+            $this->addFlash('error', 'Vous n`avez pas le droit de faire cela !');
         }
 
         return $this->redirectToRoute('app_ville_index', [], Response::HTTP_SEE_OTHER);
